trying to debug why web server keeps crashing

+++ DatabaseScripts/20211012_02_create_patients_table.sql
@@ -8,7 +8,7 @@ Create the patients table
 */
 
 CREATE TABLE [dbo].[Patients]
-	(		Id				bigint			PRIMARY KEY		IDENTITY(0, 1)
+	(		Id				int				PRIMARY KEY		IDENTITY(0, 1)
 	,		FirstName		varchar(10)		NOT NULL
 	,		LastName		varchar(20)		NOT NULL
 	,		Birthday		date			NOT NULL
