+++ RT1.Business/IPatientsService.cs
@@ -9,10 +9,10 @@ namespace RT1.Business
     public interface IPatientsService
     {
         Task<OperationOutcome<IEnumerable<PatientDto>>> GetAllPatients();
-        Task<OperationOutcome<PatientDto>> GetPatientById(long id);
+        Task<OperationOutcome<PatientDto>> GetPatientById(int id);
         Task<OperationOutcome<IEnumerable<PatientDto>>> CreateManyPatients(IEnumerable<PatientDto> patients);
         Task<OperationOutcome<IEnumerable<PatientDto>>> UpdateManyPatients(IEnumerable<PatientDto> patients);
-        Task<OperationOutcome<PatientDto>> UpdatePatient(long id, PatientDto patient);
-        Task<OperationOutcome<PatientDto>> DeletePatient(long id);
+        Task<OperationOutcome<PatientDto>> UpdatePatient(int id, PatientDto patient);
+        Task<OperationOutcome<PatientDto>> DeletePatient(int id);
     }
 }
