+++ RT1.DataProviders/IPatientsDataProvider.cs
@@ -8,10 +8,10 @@ namespace RT1.DataProviders
     public interface IPatientsDataProvider
     {
         Task<IEnumerable<Patient>> GetAllPatients();
-        Task<Patient> GetPatientById(long id);
+        Task<Patient> GetPatientById(int id);
         Task<IEnumerable<Patient>> CreateManyPatients(IEnumerable<Patient> patients);
         Task<IEnumerable<Patient>> UpdateManyPatients(IEnumerable<Patient> patients);
         Task<Patient> UpdatePatient(Patient patient);
-        Task<Patient> DeletePatient(long id);
+        Task<Patient> DeletePatient(int id);
     }
 }
